fix: removed the loop over groups for GetThreadGroupAffinity
The old version was getting the group affinity of the processor but then was applying it to every group and returning the cpu list. This led almost every processor to be added to the CPU list. Instead it should only add the current group affinity to the list which would then merged globally with an mpi routine afterwards.