feat: implemented a better vector permutation/distribution method than SpMV
initial implementation relied on the already implemented matrix-vector multiplication to perform vector permutations for the distribution. This proved to be a quite expensive operation, creating a bottleneck in the simulations. The new method with the introduction of a class designed for the sole purpose now handles this operation cheaper and more efficient, without relying on any matrix.